pyproject.toml
[tool.poetry]
name = "zoomanagerit"
version = "0.1.0"
description = ""
authors = ["fed0r <wisp47344@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
passlib = "^1.7.4"
uvicorn = "^0.34.0"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
